+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 ## [5.16.4] 2025-01-22
 
 - Doc: Exclude not relevant md from search
+- Upgrade npm dependencies
 
 ## [5.16.3] 2025-01-22